// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      restaurants: {
        Row: {
          bookability: number | null;
          created_at: string;
          distance: number | null;
          food_quality: number | null;
          google_maps_link: string | null;
          has_vegetarian_options: boolean | null;
          id: number;
          location: number | null;
          menu_link: string | null;
          name: string;
          plentiness: number | null;
          price: number | null;
          service_quality: number | null;
          time_back_and_forth: number | null;
          variety: number | null;
        };
        Insert: {
          bookability?: number | null;
          created_at?: string;
          distance?: number | null;
          food_quality?: number | null;
          google_maps_link?: string | null;
          has_vegetarian_options?: boolean | null;
          id?: number;
          location?: number | null;
          menu_link?: string | null;
          name: string;
          plentiness?: number | null;
          price?: number | null;
          service_quality?: number | null;
          time_back_and_forth?: number | null;
          variety?: number | null;
        };
        Update: {
          bookability?: number | null;
          created_at?: string;
          distance?: number | null;
          food_quality?: number | null;
          google_maps_link?: string | null;
          has_vegetarian_options?: boolean | null;
          id?: number;
          location?: number | null;
          menu_link?: string | null;
          name?: string;
          plentiness?: number | null;
          price?: number | null;
          service_quality?: number | null;
          time_back_and_forth?: number | null;
          variety?: number | null;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

export type RestaurantSchema =
  Database['public']['Tables']['restaurants']['Row'];

type PublicSchema = Database[Extract<keyof Database, 'public'>];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema['Tables'] & PublicSchema['Views'])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
        Database[PublicTableNameOrOptions['schema']]['Views'])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
      Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] &
      PublicSchema['Views'])
  ? (PublicSchema['Tables'] &
      PublicSchema['Views'])[PublicTableNameOrOptions] extends {
      Row: infer R;
    }
    ? R
    : never
  : never;
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema['Enums']
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
  ? PublicSchema['Enums'][PublicEnumNameOrOptions]
  : never;